refactor(video): wrap video methods in a class
