--- packages/restapi/src/lib/video/end.ts
@@ -1,8 +1,0 @@
-export type EndOptionsType = {
-  peerInstanceRef: any;
-};
-
-export const end = (options: EndOptionsType): void => {
-  const { peerInstanceRef } = options || {};
-  peerInstanceRef.current.destroy();
-};
