+++ packages/restapi/src/lib/video/index.ts
@@ -1,2 +1,1 @@
-export * from './create';
-export * from './request';
+export * from './Video';
